Blockquote: fix after merge
